// src/Components/Enums/FieldType.ts
export enum FieldType {
    Undefined = 'Undefined',
    Text = 'Text',
    Enum = 'Enum',
    Date = 'Date',
    Duration = 'Duration',
    Number = 'Number',
    Boolean = 'Boolean',
    Account = 'Account',
    Reference = 'Reference',
    Chart = 'Chart',
    Tile = 'Tile',
    VerticalLayout = 'VerticalLayout',
    StaticContent = 'StaticContent',
    ActionButton = 'ActionButton',
    Showcase = 'Showcase',
    Icon = 'Icon',
    CarouselLayout = 'CarouselLayout',
    Image = 'Image',
    Timeline = 'Timeline',
    Uri = "Uri",
    Table = 'Table'
};